refactor(html): extract inject-html surgery
