+++ src/cli/commands/inject.ts
@@ -2,7 +2,8 @@ import { blue, green, red } from '#cli/colors';
 import { pathFlag } from '#cli/flags/path';
 import { sizesFlag } from '#cli/flags/sizes';
 import { toDataUri } from '#dataUri';
-import { buildFaviconTags, type EmbedResolver, injectTagsIntoHtml } from '#html';
+import { buildFaviconTags, type EmbedResolver } from '#html';
+import { injectTagsIntoHtml } from '#injectHtml';
 import { DATA_URI_ENCODINGS, INJECT_MODES } from '#types';
 import { arg, CLIError, command, flag } from '@kjanat/dreamcli';
 import { mkdir, readFile, writeFile } from 'node:fs/promises';
